Keep my branch up-to-date

+++ funcs.js
@@ -19,31 +19,50 @@ module.exports = {
             if (msg.content.toLowerCase().includes(word)) {
                 switch (word) {
                     case "ban":
-                        banUser(word, msg, client);
+                        if(hasPermission(msg.member,"BAN_MEMBERS")){
+                            banUser(word, msg, client);
+                        }
                         break;
 
                     case "kick":
-                        kickUser(word, msg, client);
+                        if(hasPermission(msg.member,"KICK_MEMBERS")){
+                            kickUser(word, msg, client);
+                        }
                         break;
 
                     case "timeout":
-                        timeoutUser(word, msg, client);
+                    
+                        if(hasPermission(msg.member,"MANAGE_MESSAGES")){
+                            timeoutUser(word, msg, client);
+                        }
                         break;
 
                     case "enable" || "activate":
-                        featureToggle(msg, true);
+                        if(hasPermission(msg.member,"ADMINISTRATOR")){
+                            var type = true;
+                            featureToggle(msg, type);
+                        }
                         break;
 
                     case "disable" || "deactivate":
-                        featureToggle(msg, false);
+                        
+                        if(hasPermission(msg.member,"ADMINISTRATOR")){
+                            var type = false;
+                            featureToggle(msg, type);
+                        }
                         break;
 
                     case "turn":
-                        featureToggle(msg, null);
+                        if(hasPermission(msg.member,"ADMINISTRATOR")){
+                            featureToggle(msg, null);
+                        }
                         break;
 
                     case "purge":
-                        purgeChannel(msg, client)
+                        if(hasPermission(msg.member,"MANAGE_MESSAGES")){
+                            purgeChannel(msg, client);
+                            
+                        }
                         break;
                 }
             }
@@ -54,7 +73,7 @@ module.exports = {
             for (const value of msg.attachments.array().values()) {
                 userinfoattachmentstack.push({ attachments: [{ filename: value.filename, url: value.url }] });
             }
-            userinfostack.push({ name: msg.author.username, id: msg.author.id, msg: msg.content, msgattachments: userinfoattachmentstack, msgId: msg.id, createdAt: msg.createdAt });
+            userinfostack.push({ name: msg.author.username,channel: msg.channel.id, id: msg.author.id, msg: msg.content, msgattachments: userinfoattachmentstack, msgId: msg.id, createdAt: msg.createdAt });
             userinfoattachmentstack = [];
         }
         else {
@@ -271,11 +290,13 @@ function timeoutUser(word, msg, client) {
 
         ids.forEach(function (id) {
             if (id != "592783579998584868") {
-                end = calculateTimeoutEnd(id, timer);
-                if (end != 0) {
-                    var member = msg.guild.members.get(id);
-                    timeouts.push(member.displayName);
-                    member.send("You have been timed out from **" + client.guilds.get(msg.guild.id).name + "**, you will be able to chat again in **" + end + "**");
+                if(isTimedout(id) == false){
+                    end = calculateTimeoutEnd(id, timer);
+                    if (end != 0) {
+                        var member = msg.guild.members.get(id);
+                        timeouts.push(member.displayName);
+                        member.send("You have been timed out from **" + client.guilds.get(msg.guild.id).name + "**, you will be able to chat again in **" + end + "**");
+                    }
                 }
 
             }
@@ -321,7 +342,7 @@ function calculateTimeoutEnd(did, times) {
         });
     }
 
-    if (m != 0) {
+    if (s != 0) {
         times.seconds.forEach(function (seconds) {
             suc = suc + parseInt(seconds.replace(/\D/g, ''));
         });
@@ -415,89 +436,32 @@ function translateEmbed(msg, translation) {
 }
 
 function featureToggle(msg, toggleType) {
-
+    console.log(toggleType);
     var types = ["translation", "nsfw", "cooldown", "blacklisting"]
 
-    if (msg.content.toLowerCase().includes("off")) {
+    if (msg.content.toLowerCase().includes(" off ")) {
         toggleType = false;
-    } else if (msg.content.toLowerCase().includes("on")) {
+    } else if (msg.content.toLowerCase().includes(" on ")) {
         toggleType = true;
-    } else if (msg.content.toLowerCase().includes("off") && msg.content.toLowerCase().includes("on")) {
+    } else if (msg.content.toLowerCase().includes(" off ") && msg.content.toLowerCase().includes(" on ")) {
         toggleType = null;
+    }else{
+        toggleType = toggleType;
     }
 
     if (toggleType == null) {
-        msg.channel.send(msg.author + ", It appears you are trying to enabled and disable a feature at the same time ");
+        msg.channel.send(msg.author + ", it appears you are trying to enabled and disable a feature at the same time.");
     } else {
         types.forEach(function (type) {
             if (msg.content.toLowerCase().includes(type)) {
-                var state = toggleType = true ? "Enabled." : "Disabled.";
-                switch (type) {
-                    case "translation":
-                        configFile["Translation"] = toggleType;
-                        msg.channel.send(msg.author + ",  the __Auto Translation__ feature has been " + state);
-                        break;
-
-                    case "nsfw":
-                        configFile["NSFW Filter"] = toggleType;
-                        msg.channel.send(msg.author + ",  the __NSFW Filter__ feature has been " + state);
-                        break;
-
-                    case "cooldown":
-                        configFile["Statistical Slow-Mode"] = toggleType;
-                        msg.channel.send(msg.author + ",  the __Statistical Slow-Mode__ feature has been " + state);
-                        break;
+                var state;
+                console.log(toggleType);
+                if(toggleType){
+                     state = "Enabled.";
+                }else{
+                     state = "Disabled.";
                 }
-            }
-        });
-    }
-}
 
-
-function getTranslatedText(msg, callback) {
-    var translated;
-
-    if (msg.content == "" && msg.attachments.size > 0) {
-        return;
-    }
-    else {
-        var options = {
-            method: 'GET',
-            url: 'https://translate.yandex.net/api/v1.5/tr.json/translate',
-            qs:
-            {
-                key: 'trnsl.1.1.20190626T023402Z.aec9c733ab816267.ead2c2e94b6b8e1dfe3057775ce1613d21a92e38',
-                lang: config.lang,
-                text: msg.content
-            }
-        };
-
-        request(options, function (error, response, body) {
-            if (error) throw new Error(error);
-            var obj = JSON.parse(body);
-            callback(obj.text[0]);
-        });
-    }
-}
-
-function featureToggle(msg, toggleType) {
-
-    var types = ["translation", "nsfw", "cooldown", "blacklisting"]
-
-    if (msg.content.toLowerCase().includes("off")) {
-        toggleType = false;
-    } else if (msg.content.toLowerCase().includes("on")) {
-        toggleType = true;
-    } else if (msg.content.toLowerCase().includes("off") && msg.content.toLowerCase().includes("on")) {
-        toggleType = null;
-    }
-
-    if (toggleType == null) {
-        msg.channel.send(msg.author + ", It appears you are trying to enabled and disable a feature at the same time ");
-    } else {
-        types.forEach(function (type) {
-            if (msg.content.toLowerCase().includes(type)) {
-                var state = toggleType ? "Enabled." : "Disabled.";
                 switch (type) {
                     case "translation":
                         configFile["Translation"] = toggleType;
@@ -528,10 +492,27 @@ function featureToggle(msg, toggleType) {
     }
 }
 
+
 function purgeChannel(msg, client) {
-    if (msg.mentions.members.array()[0].user.id == client.id) {
-        if (msg.mentions.members.array()[1].exists()) {
-            var userid = msg.mentions.members.array()[1].user.id;
+    if (msg.mentions.channels.size > 0) {
+        msg.mentions.channels.forEach((function(channel){
+            var channelId = channel.id;
+            client.guilds.get(msg.guild.id).channels.get(channelId).fetchMessages().then(messages => messages.array().forEach(message => {
+                message.delete();
+            }));
+        }));
+    }else if (msg.mentions.members.size > 0) {
+        if (msg.mentions.members.array()[0].user.id == client.id) {
+            if (msg.mentions.members.array()[1].exists()) {
+                var userid = msg.mentions.members.array()[1].user.id;
+                msg.channel.fetchMessages().then(messages => messages.array().forEach(message => {
+                    if (message.author.id == userid) {
+                        message.delete();
+                    }
+                }));
+            }
+        }else {
+            var userid = msg.mentions.members.array()[0].user.id;
             msg.channel.fetchMessages().then(messages => messages.array().forEach(message => {
                 if (message.author.id == userid) {
                     message.delete();
@@ -539,12 +520,29 @@ function purgeChannel(msg, client) {
             }));
         }
     }
-    else {
-        var userid = msg.mentions.members.array()[0].user.id;
-        msg.channel.fetchMessages().then(messages => messages.array().forEach(message => {
-            if (message.author.id == userid) {
-                message.delete();
-            }
-        }));
-    }
+
+    msg.channel.send(msg.author + ", Now purging");
+
+}
+
+function isTimedout(id){
+    var isTimedout = false;
+    var filedata = fs.readFileSync('./timeouts.json', { encoding: 'utf8' });
+    var timeouts = JSON.parse(filedata);
+    timeouts["active"].forEach(function (obj) {
+        if (obj.id == id) {
+            var timeEnd = new Date(obj.end);
+            var now = new Date();
+            isTimedout = true;
+        }
+    });
+    return isTimedout;
+}
+
+function hasPermission(member,permission){
+    return member.hasPermission(permission) ? true : false;
+}
+
+function statSlowmode(){
+
 }
